v3: Switched to LGSVL 2020.06 and Unity 2019.3.15f Dynamic and Static object spawn added. Ready to test swerve and stop.

+++ Jenkins/build-simulator.sh
@@ -68,8 +68,8 @@ else
   CONTROLLABLES=
 fi
 
-if [ ! -z ${SIMULATOR_SENSORS+x} ]; then
-  SENSORS="-buildBundles -buildSensors ${SIMULATOR_SENSORS}"
+if [ ! -z ${SIM_SENSORS+x} ]; then
+  SENSORS="-buildBundles -buildSensors ${SIM_SENSORS}"
 else
   SENSORS=
 fi
@@ -234,8 +234,10 @@ cp /mnt/AssetBundles/Controllables/controllable_* /tmp/${BUILD_OUTPUT}/AssetBund
 mkdir -p /tmp/${BUILD_OUTPUT}/AssetBundles/Sensors
 cp /mnt/AssetBundles/Sensors/sensor_* /tmp/${BUILD_OUTPUT}/AssetBundles/Sensors
 
-mkdir -p /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
-cp -R /mnt/AssetBundles/NPCs/* /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
+if [ ! -z ${SIMULATOR_NPCS+x} ]; then
+    mkdir -p /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
+    cp -R /mnt/AssetBundles/NPCs/* /tmp/${BUILD_OUTPUT}/AssetBundles/NPCs
+fi
 
 cd /tmp
 zip -r /mnt/${BUILD_OUTPUT}.zip ${BUILD_OUTPUT}
